Fix seeder keys

+++ database/migrations/2020_12_09_172207_links.php
@@ -17,7 +17,7 @@ class Links extends Migration
             $table->bigIncrements('id');
 
             // User
-            $table->mediumInteger('user_id')->unsigned()->nullable(false)->index();
+            $table->bigInteger('user_id')->unsigned()->nullable(false)->index();
             $table->foreign('user_id')->references('id')->on('users');
 
             // URL
